moved files into appropriate folders and got rid of empty picture list item
